Make Easymotion keys visible
The current readme uses tag delimiters which results in part of the readme being hidden from the end user. It is also not obvious that the commas are keys and not just a break in the sentence. For this and other reasons I also wrapped all keys in the kdb tags were are specifically intended for this situation.